refactor(agent): 重构多个agent类继承自BaseAgent
将browser_use_agent、deep_analyzer_agent、deep_researcher_agent和planning_agent重构为继承自新创建的BaseAgent类，减少代码重复。BaseAgent类包含了这些agent共有的核心功能实现。

主要变更：
1. 创建BaseAgent基类包含共享逻辑
2. 简化各子类实现，仅保留特定配置
3. 统一初始化流程
4. 添加.gitignore忽略.idea目录
